fixed a warning
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@603 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/ANALYSIS/SVM/SVMWrapper.C
@@ -124,7 +124,7 @@ namespace OpenMS
 	    switch(type)
 	    {
 				case(DEGREE):
-					param_->degree = value;
+					param_->degree = (int)value;
 				  break;
 				case(C):
 			    param_->C = value;
